fusion-engine/requirements.txt
aiohttp==3.9.0
websockets==12.0
redis==5.0.0
numpy==1.24.0
scipy==1.11.0
filterpy==1.4.2
python-dotenv==1.0.0
pydantic==2.5.0
